requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
pydantic==2.5.2
python-multipart==0.0.6
pdfplumber==0.10.3
openai==1.3.5
python-dotenv==1.0.0
google-generativeai==0.4.0
